rename ErrorResposne to TxErrorResponse and locate it to components
